Move rpcfs/ to termite/
